+++ src/main/kotlin/Application.kt
@@ -1,6 +1,7 @@
 package com.cubis
 
 import com.cubis.jasper.FontRegistry
+import com.cubis.jasper.TemplateManager
 import com.cubis.plugins.configureSerialization
 import io.ktor.server.application.*
 
@@ -11,6 +12,9 @@ fun main(args: Array<String>) {
 fun Application.module() {
     // Initialize font registry on startup
     FontRegistry.init()
+    
+    // Initialize template manager on startup
+    TemplateManager.init()
     
     configureSerialization()
     configureRouting()