+++ src/main/kotlin/Routing.kt
@@ -2,6 +2,7 @@ package com.cubis
 
 import com.cubis.jasper.fontRoutes
 import com.cubis.jasper.jasperReportRoutes
+import com.cubis.jasper.templateRoutes
 import io.ktor.server.application.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -14,5 +15,6 @@ fun Application.configureRouting() {
         
         jasperReportRoutes()
         fontRoutes()
+        templateRoutes()
     }
 }